+++ tests/testthat/test-random_effects_ci-glmmTMB.R
@@ -14,11 +14,9 @@ osx <- tryCatch(
   }
 )
 
-## TODO: Need fixes for glmmTMB
 ## TODO also check messages for profiled CI
 
-if (FALSE &&
-    .runThisTest && !osx &&
+if (.runThisTest && !osx &&
     requiet("testthat") &&
     requiet("parameters") &&
     requiet("glmmTMB") &&
@@ -29,27 +27,25 @@ if (FALSE &&
   set.seed(123)
   sleepstudy$Months <- sample(1:4, nrow(sleepstudy), TRUE)
 
-  m1 <- glmmTMB(angle ~ temperature + (temperature | recipe) + (temperature | replicate), data = cake)
+  m1 <- suppressWarnings(glmmTMB(angle ~ temperature + (temperature | recipe) + (temperature | replicate), data = cake))
   m2 <- glmmTMB(Reaction ~ Days + (Days | Subject), data = sleepstudy)
-  m3 <- glmmTMB(angle ~ temperature + (temperature | recipe), data = cake)
-  m4 <- glmmTMB(angle ~ temperature + (temperature | replicate), data = cake)
-  m5 <- glmmTMB(Reaction ~ Days + (Days + Months | Subject), data = sleepstudy)
+  m3 <- suppressWarnings(glmmTMB(angle ~ temperature + (temperature | recipe), data = cake))
+  m4 <- suppressWarnings(glmmTMB(angle ~ temperature + (temperature | replicate), data = cake))
+  m5 <- suppressWarnings(glmmTMB(Reaction ~ Days + (Days + Months | Subject), data = sleepstudy))
 
-  mp1 <- model_parameters(m1) # fails
+  expect_message(expect_message(mp1 <- model_parameters(m1), "singularity"), "singularity")
   mp2 <- model_parameters(m2) # works
-  mp3 <- model_parameters(m3) # no SE/CI
-  mp4 <- model_parameters(m4) # no SE/CI
-  mp5 <- model_parameters(m5) # no SE/CI
+  expect_message(mp3 <- model_parameters(m3), "singularity") # no SE/CI
+  expect_message(mp4 <- model_parameters(m4), "singularity") # no SE/CI
+  expect_message(mp5 <- model_parameters(m5), "singularity") # no SE/CI
 
   test_that("random effects CIs, two slopes, categorical", {
     expect_equal(
       mp1$CI_low,
-      c(28.75568, 4.97893, -1.95002, -2.69995, -3.62201, -2.69102,
-        4.28558, 0.21474, 0.40062, 0.10169, 0.04953, 1e-05, 0.55398,
-        0, 2e-05, 0.6333, 1.09851, 0.00944, -1, -1, -0.65406, -1, -0.69103,
-        -1, -0.95271, -1, -0.90617, -1, -0.99802, -0.99836, -1, -1, -1,
-        -1, -0.75274, -0.96895, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-        -1, -1, 4.07985),
+      c(28.9123, 5.03115, -1.87304, -2.42081, -3.2708, -2.57695, NA,
+        NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
+        NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
+        NA, NA, NA, NA, NA, NA, NA, NA, NA, NA),
       tolerance = 1e-3,
       ignore_attr = TRUE
     )
@@ -61,22 +57,22 @@ if (FALSE &&
         "SD (temperature.L)", "SD (temperature.Q)", "SD (temperature.C)",
         "SD (temperature^4)", "SD (temperature^5)", "SD (temperature.L)",
         "SD (temperature.Q)", "SD (temperature.C)", "SD (temperature^4)",
-        "SD (temperature^5)", "Cor (Intercept~temperature.C: replicate)",
-        "Cor (Intercept~temperature.C: recipe)", "Cor (Intercept~temperature.L: replicate)",
-        "Cor (Intercept~temperature.L: recipe)", "Cor (Intercept~temperature.Q: replicate)",
-        "Cor (Intercept~temperature.Q: recipe)", "Cor (Intercept~temperature^4: replicate)",
-        "Cor (Intercept~temperature^4: recipe)", "Cor (Intercept~temperature^5: replicate)",
-        "Cor (Intercept~temperature^5: recipe)", "Cor (temperature.L~temperature.C: replicate)",
-        "Cor (temperature.Q~temperature.C: replicate)", "Cor (temperature.L~temperature.Q: replicate)",
-        "Cor (temperature.L~temperature^4: replicate)", "Cor (temperature.Q~temperature^4: replicate)",
-        "Cor (temperature.C~temperature^4: replicate)", "Cor (temperature.L~temperature^5: replicate)",
-        "Cor (temperature.Q~temperature^5: replicate)", "Cor (temperature.C~temperature^5: replicate)",
-        "Cor (temperature^4~temperature^5: replicate)", "Cor (temperature.L~temperature.C: recipe)",
+        "SD (temperature^5)", "Cor (Intercept~temperature.C: recipe)",
+        "Cor (Intercept~temperature.C: replicate)", "Cor (Intercept~temperature.L: recipe)",
+        "Cor (Intercept~temperature.L: replicate)", "Cor (Intercept~temperature.Q: recipe)",
+        "Cor (Intercept~temperature.Q: replicate)", "Cor (Intercept~temperature^4: recipe)",
+        "Cor (Intercept~temperature^4: replicate)", "Cor (Intercept~temperature^5: recipe)",
+        "Cor (Intercept~temperature^5: replicate)", "Cor (temperature.L~temperature.C: recipe)",
         "Cor (temperature.Q~temperature.C: recipe)", "Cor (temperature.L~temperature.Q: recipe)",
         "Cor (temperature.L~temperature^4: recipe)", "Cor (temperature.Q~temperature^4: recipe)",
         "Cor (temperature.C~temperature^4: recipe)", "Cor (temperature.L~temperature^5: recipe)",
         "Cor (temperature.Q~temperature^5: recipe)", "Cor (temperature.C~temperature^5: recipe)",
-        "Cor (temperature^4~temperature^5: recipe)", "SD (Observations)"
+        "Cor (temperature^4~temperature^5: recipe)", "Cor (temperature.L~temperature.C: replicate)",
+        "Cor (temperature.Q~temperature.C: replicate)", "Cor (temperature.L~temperature.Q: replicate)",
+        "Cor (temperature.L~temperature^4: replicate)", "Cor (temperature.Q~temperature^4: replicate)",
+        "Cor (temperature.C~temperature^4: replicate)", "Cor (temperature.L~temperature^5: replicate)",
+        "Cor (temperature.Q~temperature^5: replicate)", "Cor (temperature.C~temperature^5: replicate)",
+        "Cor (temperature^4~temperature^5: replicate)", "SD (Observations)"
       )
     )
   })
@@ -85,7 +81,7 @@ if (FALSE &&
   test_that("random effects CIs, simple slope", {
     expect_equal(
       mp2$CI_low,
-      c(237.93546, 7.41637, 15.5817, 3.91828, -0.50907, 22.80044),
+      c(238.40611, 7.52295, 15.01709, 3.80546, -0.48781, 22.80047),
       tolerance = 1e-3,
       ignore_attr = TRUE
     )
@@ -101,9 +97,9 @@ if (FALSE &&
   test_that("random effects CIs, categorical slope-1", {
     expect_equal(
       mp3$CI_low,
-      c(30.91139, 4.33247, -2.6798, -3.20703, -4.07681, -3.27237, 0.06301,
-        0, 0, 0.1192, 0.32213, 0, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-        -1, -1, -1, -1, -1, -1, 7.09933),
+      c(31.20278, 4.35879, -2.63767, -2.80041, -3.54983, -3.16627,
+        NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
+        NA, NA, NA, NA, NA, NA),
       tolerance = 1e-3,
       ignore_attr = TRUE
     )
@@ -129,10 +125,9 @@ if (FALSE &&
   test_that("random effects CIs, categorical slope-2", {
     expect_equal(
       mp4$CI_low,
-      c(28.88523, 4.96796, -1.93239, -1.98597, -2.68858, -2.5524, 4.27899,
-        0.35378, 0.08109, 0.03419, 0, 0.49982, -0.68893, -0.71984, -1,
-        -0.96725, -0.92158, -0.99894, -0.99924, -1, -1, -1, -1, -0.80378,
-        -0.9778, -1, -1, 4.21143),
+      c(29.01131, 5.01247, -1.89444, -1.96275, -2.66798, -2.50892,
+        NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
+        NA, NA, NA, NA, NA, NA),
       tolerance = 1e-3,
       ignore_attr = TRUE
     )
@@ -157,7 +152,7 @@ if (FALSE &&
   test_that("random effects CIs, double slope", {
     expect_equal(
       mp5$CI_low,
-      c(237.99863, 7.4022, 12.63814, 0.58664, 0, -0.58599, -1, -1, 22.65226),
+      c(238.40607, 7.52296, NA, NA, NA, NA, NA, NA, NA),
       tolerance = 1e-3,
       ignore_attr = TRUE
     )
@@ -180,15 +175,15 @@ if (FALSE &&
   sleepstudy$Months <- sample(1:4, nrow(sleepstudy), TRUE)
 
   m2 <- glmmTMB(Reaction ~ Days + (0 + Days | Subject), data = sleepstudy)
-  m5 <- glmmTMB(Reaction ~ Days + (0 + Days + Months | Subject), data = sleepstudy)
+  m5 <- suppressWarnings(glmmTMB(Reaction ~ Days + (0 + Days + Months | Subject), data = sleepstudy))
 
   mp2 <- model_parameters(m2)
-  mp5 <- model_parameters(m5)
+  expect_message(mp5 <- model_parameters(m5), "singularity") # no SE/CI
 
   test_that("random effects CIs, simple slope", {
     expect_equal(
       mp2$CI_low,
-      c(243.47155, 6.77765, 5.09041, 26.01525),
+      c(243.55046, 6.89554, 4.98429, 25.94359),
       tolerance = 1e-3,
       ignore_attr = TRUE
     )
@@ -202,7 +197,7 @@ if (FALSE &&
   test_that("random effects CIs, simple slope", {
     expect_equal(
       mp5$CI_low,
-      c(241.61021, 7.43503, 4.11446, 2.69857, -0.40595, 24.632),
+      c(237.03695, 9.04139, NA, NA, NA, NA),
       tolerance = 1e-3,
       ignore_attr = TRUE
     )
@@ -212,6 +207,13 @@ if (FALSE &&
       c("(Intercept)", "Days", "SD (Days)", "SD (Months)", "Cor (Days~Months: Subject)",
         "SD (Observations)")
     )
+  })
+
+
+  # messages for profiled CI
+  test_that("profiled CI messages", {
+    mp2 <- model_parameters(m2, ci_method = "profile")
+    expect_message(print(mp2), regexp = "(.*)profile-likelihood(.*)z-distribution(.*)")
   })
 
 }